@@ -1,7 +1,7 @@
 [package]
 name = "hyper-server"
 version = "0.0.3"
-edition = "2018"
+edition = "2024"
 authors = []
 
 [[bin]]
@@ -9,14 +9,13 @@ name = "hyper-server"
 path = "src/main.rs"
 
 [dependencies]
-arc-swap = "1.7.1"
 futures = "0.3"
 http-body-util = "0.1.1"
-hyper = { version = "1", features = ["full"] }
-hyper-util = { version = "0.1.5", features = ["tokio"] }
+hyper = { version = "1.6", features = ["full"] }
+hyper-util = { version = "0.1.14", features = ["tokio", "server", "server-auto"] }
 libc = "0.2"
 log = "0.4"
-ppp = "2.2.0"
+ppp = "2.3"
 tokio = {"version"= "1", features=["rt-multi-thread", "macros", "net"]}
 tracing = "0.1"
 tracing-subscriber = "0.3"
